Revert "move more stuff"
This reverts commit 022caf66749f427b28e885a5b3188267ec3188cd.
