+++ napari/utils/__init__.py
@@ -1,4 +1,4 @@
-from ._nbscreenshot import nbscreenshot
+from .._qt.utils.nbscreenshot import NotebookScreenshot as nbscreenshot
 from .colormaps import Colormap
 from .dask_utils import resize_dask_cache
 from .info import citation_text, sys_info
